Jwt completed.

+++ Msdi.Authentication/Models/TokenOptions.cs
@@ -8,7 +8,7 @@ namespace Msdi.Authentication.Models
     {
         public string Audience { get; set; }
         public string Issuer { get; set; }
-        public int AccessTokenExpiration { get; set; }
+        public double AccessTokenExpiration { get; set; }
         public string SecurityKey { get; set; }
     }
 }
